(IK-26) Add new component , add new file constant
