+++ src/App.js
@@ -5,6 +5,7 @@ import Header from "./components/Header/Header";
 import Profile from "./components/Profile/Profile";
 import Dialogs from "./components/Dialogs/Dialogs";
 import { BrowserRouter, Route } from "react-router-dom";
+import { DIALOGS, PROFILE } from "./components/NavBar/constats";
 
 const App = (props) => {
   return (
@@ -13,8 +14,8 @@ const App = (props) => {
         <Header />
         <Nav />
         <div class="app-wrapper-content">
-          <Route path="/profile" component={Profile} />
-          <Route path="/dialogs" component={Dialogs} />
+          <Route path={PROFILE} component={Profile} />
+          <Route path={DIALOGS} component={Dialogs} />
         </div>
       </div>
     </BrowserRouter>
