+++ src/components/Dialogs/Dialogs.jsx
@@ -1,22 +1,7 @@
 import React from "react";
 import s from "./Dialogs.module.css";
-import { NavLink } from "react-router-dom";
-
-const DialogItem = (props) => {
-  let path = "/dialogs/" + props.id;
-
-  return (
-    <div className={s.dialog}>
-      <NavLink to={path} activeClassName={s.active}>
-        {props.name}
-      </NavLink>
-    </div>
-  );
-};
-
-const Message = (props) => {
-  return <div className={s.dialog}>{props.message}</div>;
-};
+import DialogItem from "./DialogItem/DialogsItem";
+import Message from "./Message/Message";
 
 const dialogs = [
   { id: 1, name: "Ruslana" },
