+++ src/components/Dialogs/Message/Message.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import s from "./../Dialogs.module.css";
+
+const Message = (props) => {
+  return <div className={s.dialog}>{props.message}</div>;
+};
+
+export default Message;
