+++ src/components/NavBar/constats.js
@@ -1,0 +1,5 @@
+export const DIALOGS = "/dialogs";
+export const PROFILE = "/profile";
+export const NEWS = "/news";
+export const MUSIC = "/music";
+export const SETTING = "/settings";
